requirements.txt
numpy
numpy-stl
manim
matplotlib
